Add end of deck handling when dealing cards
deal_cards() now handles when end of deck occurs mid deal. This is accomplished
by adding a try and except statement that attempts to deal a card but resets
the table shoe on IndexError. deal_cards() now recursively calls itself until
all players and the dealer have the necessary 2 cards to play.