Add universal handling of end of deck for both dealing and player hits
Create get_top_card method to Shoe class. This method will return the top card
of the shoe if there is at least one card left. Otherwise the shoe will be reset
on the the reset method, and get_top_card method is called recursively. This new
method should be used for any card dealt from the deck to ensure end of deck
is handled properly.