Add generate_summary method to Hand class, handling for Aces
generate_summary method of class Hand checks if there are any big Aces (11 pts)
present in the hand. If so, it alters the hand summary to show alternative
score if the Ace is to be worth 1 pt. The check_for_bust method has been
updated to check for big Aces and downgrade them to 1 pt if the total
is over 21. This means, the user does not have to explicitly choose 1 vs 11 pts
when they have an Ace in their hand.